Attempt to make zmnebi load faster
Added font-display: swap to all the CSS so that the page won't be blank if the fonts take a bit to load. Also, updated the TOC to hopefully be a bit clearer. I tried using position: sticky on the headers, but due to the variable size of hte header names, it doesn't work too well.